% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annulus2d}
\alias{annulus2d}
\title{2-dimensional point cloud of an annulus}
\format{
a matrix with 100 rows and 2 columns: the x- and y-coordinates
}
\source{
\url{https://github.com/rrrlw/ggtda/blob/master/data-raw/annulus2d.R}
}
\usage{
annulus2d
}
\description{
A dataset containing the Cartesian coordinates of 100 points selected
from a 2-dimensional annulus (noisy unit circle).
}
\keyword{datasets}
