% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{dgm_select}
\alias{dgm_select}
\title{Persistence diagram selection}
\usage{
dgm_select(input, output, session)
}
\arguments{
\item{input, }{output, session standard \code{shiny} boilerplate}
}
\value{
list with following components
\describe{
\item{plot_type}{reactive character string indicating the type of diagram to plot}
}
}
\description{
Persistence diagram selection
}
