% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{dgm_var}
\alias{dgm_var}
\title{dgm_var
Variables selection sever-module for persistence diagrams.}
\usage{
dgm_var(input, output, session, dgm)
}
\arguments{
\item{dgm}{a (non-reactive) data.frame with names 'appear', 'disappear', and 'dim'.}
}
\value{
list with following components
\describe{
\item{time_window}{reactive numeric vector indicating the beirth and death time window to show}
\item{dims_selected}{reactive integer indicating the homology groups to show}
\item{noise_threshold}{reactive character string indicating the noise threshold}
}
}
\description{
dgm_var
Variables selection sever-module for persistence diagrams.
}
