% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{dmg_select_ui}
\alias{dmg_select_ui}
\title{dmg_select_ui}
\usage{
dmg_select_ui(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Shiny UI module for interacting with various types of persistence data.
}
