% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist.R
\name{geom_persist}
\alias{geom_persist}
\title{Plot Persistent Homology as a Persistence Diagram}
\usage{
geom_persist(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{stat}{The statistical transformation used to display the data.
Defaults to `identity`; another useful option is `flat`.
Pass a string to override the default.}

\item{...}{additional arguments passed to [ggplot2::layer()]}
}
\description{
Plots a flat persistence diagram
}
\seealso{
[ggplot2::layer()] for additional arguments.

Other TDA plot layers: \code{\link{stat_flat}}
}
\concept{TDA plot layers}
