% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtda.R
\docType{package}
\name{ggtda}
\alias{ggtda}
\title{A \emph{ggplot2} Extension for Persistence Data}
\description{
\emph{ggtda} is a user-friendly \emph{ggplot2} extension for visualizing data arising
from topological data analysis, including persistence data and topological
constructions on planar data.
}
