% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module.R
\name{interact_phom}
\alias{interact_phom}
\title{Interactive Exploration of Dataset Topology using Shiny}
\usage{
interact_phom(dataset)
}
\arguments{
\item{dataset}{two-column data frame or tibble}
}
\value{
shiny object
}
\description{
Allows user to change simplicial complex radius for a 2-dimensional
point cloud and observe dynamic changes in simplicial complex structure.
}
