% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{linkedPersistence}
\alias{linkedPersistence}
\title{linkedPersistence}
\usage{
linkedPersistence(input, output, session, dgm, dgm_selection,
  dgm_vars = reactive({     NULL }))
}
\arguments{
\item{dgm}{a (non-reactive) data.frame with names 'birth', 'death', and 'dimension'.}

\item{dgm_vars}{list containing reactive expressions 'time_window', 'dims_selected', and 'noise_threshold'}

\item{plot_params}{reactive expression of ggplot2 elements to add}
}
\value{
list with following components
\describe{
\item{xvar}{reactive character string indicating x variable selection}
\item{yvar}{reactive character string indicating y variable selection}
}
}
\description{
linkedPersistence
}
