% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{linkedPersistenceOutput}
\alias{linkedPersistenceOutput}
\title{linkedPersistenceOutput}
\usage{
linkedPersistenceOutput(id, ...)
}
\arguments{
\item{id}{}
}
\description{
Output UI module for creating a plot of persistence data.
}
