% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplicial-complex.R
\name{simplicial_complex}
\alias{simplicial_complex}
\alias{stat_simplicial_complex}
\alias{geom_simplicial_complex}
\title{Simplicial complexes from 2-D point clouds}
\usage{
stat_simplicial_complex(
  mapping = NULL,
  data = NULL,
  geom = "SimplicialComplex",
  position = "identity",
  complex = "Rips",
  diameter = NULL,
  radius = NULL,
  dimension_max = 2L,
  zero_simplices = "all",
  one_simplices = "maximal",
  engine = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

geom_simplicial_complex(
  mapping = NULL,
  data = NULL,
  stat = "SimplicialComplex",
  position = "identity",
  outlines = TRUE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[ggplot2:layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{complex}{The type of complex to compute, one of \code{"Vietoris"}, \code{"Rips"}
(equivalent), \code{"Cech"}, or \code{"alpha"}.}

\item{radius, diameter}{The (positive) radius or diameter used in the
construction. Provide only one of these; if neither is provided, they
default to zero.}

\item{dimension_max}{Compute simplices of dimension up to \code{dimension_max}
(only relevant for the Vietoris--Rips complex computed with the
\code{simplextree} engine).}

\item{zero_simplices}{Which 0-simplices (vertices) to plot; one of \code{"none"},
\code{"maximal"}, and \code{"all"} (default).}

\item{one_simplices}{Which 1-simplices (edges) to plot; one of \code{"none"},
\code{"maximal"} (default), and \code{"all"}.}

\item{engine}{The computational engine to use (see 'Details'). Reasonable
defaults are chosen based on \code{complex}.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{outlines}{Should the outlines of polygons representing high-dimensional
simplices be drawn?}
}
\description{
Construct and plot simplicial complexes that equal or
approximate the topology of a ball cover of a set of points.
}
\details{
Persistent homology is ultimately based on the topological properties of
regions containing a set of points. When the region is the union of balls of
a common radius, its homology is equal to or approximated by that of several
families of \emph{simplicial complexes} constructed on the point set. The
simplicial complex stat constructs these simplicial complexes for a set of
points in \eqn{xy}-space while the geom plots them on the same coordinates as
the points.
}
\section{Complexes}{

A \emph{Vietoris--Rips complex} of a point cloud is the simplicial complex
consisting of a simplex for each subset of points within a fixed diameter
of each other. A \emph{Čech complex} contains the simplex for each subset that
lies within a circle of fixed diameter. (This means that the Čech complex
depends on the geometry of the ambient space containing the point cloud,
while the Vietoris complex depends only on the inter-point distances.
Moreover, a Vietoris complex contains the Čech complex of the same
diameter.) An \emph{alpha complex} comprises those simplices of the Delaunay
triangulation within a fixed diameter.

{\strong{ggtda}} relies on four engines to compute simplicial complexes, which
can be specified to the \code{engine} parameter: Vietoris--Rips and Čech
complexes of dimension at most 2 are implemented in base R (\code{"base"}),
which is slow but allows the package to stand alone for small cases.
\code{\link[RTriangle:triangulate]{RTriangle::triangulate()}} is used to compute the Delaunay triangulation
for alpha complexes (\code{"RTriangle"}), without inserting Steiner points (so
that the vertices of the triangulation are among those of the data). The
package \strong{TDA} can compute \link[TDA:ripsFiltration]{Vietoris--Rips filtrations} and \link[TDA:alphaComplexFiltration]{alpha filtrations} (\code{"TDA"} for default engines,
or specify the \code{"GUDHI"} or \code{"Dionysus"} engine). Finally, the highly
optimized package
\strong{\link[simplextree:simplextree-package]{simplextree}} can be called
to compute Vietoris--Rips complexes (\code{"simplextree"}). As other complexes
are implemented in {\strong{simplextree}}, they will be made available here.
}

\section{Aesthetics}{

\code{stat_simplicial_complex()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \code{group}
}
\code{geom_simplicial_complex()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \strong{\code{id}}
\item \strong{\code{dimension}}
\item \code{alpha}
\item \code{colour}
\item \code{fill}
\item \code{group}
\item \code{linetype}
\item \code{linewidth}
\item \code{shape}
\item \code{size}
\item \code{stroke}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs", package = "ggplot2")}.
}

\section{Computed variables}{

These are calculated by the 'stat' part of layers and can be accessed with \link[ggplot2:aes_eval]{delayed evaluation}.
\itemize{
\item \code{after_stat(dimension)}\cr integer dimension of the corresponding simplex.
\item \code{after_stat(id)}\cr character simplex identifier within each \code{dimension}.
\item \code{after_stat(face)}\cr factor encoding of \code{dimension} for \verb{>= 2L}-dimensional simplices.
}
}

\examples{

set.seed(1)
s <- seq(0, 2*pi, length.out = 40)
df <- data.frame(
  x = cos(s) + rnorm(40, 0, .1),
  y = sin(s) + rnorm(40, 0, .1)
)

# default
ggplot(df, aes(x, y)) +
  coord_fixed() +
  geom_simplicial_complex(radius = .4)
# higher-dimensional simplices are not more opaque but cause overlap
ggplot(df, aes(x, y)) +
  coord_fixed() +
  geom_simplicial_complex(radius = .4, dimension_max = 3L)

# visualizing dimension w/ face & alpha:
ggplot(df, aes(x, y)) +
  coord_fixed() +
  geom_simplicial_complex(
    mapping = aes(alpha = after_stat(face)),
    radius = .3, dimension_max = 4L
  ) +
  scale_alpha_ordinal(range = c(.1, .6))
# visualizing dimension w/ dimension & fill:
ggplot(df, aes(x, y)) +
  coord_fixed() +
  geom_simplicial_complex(
    mapping = aes(fill = after_stat(dimension)),
    radius = .3, dimension_max = 4L, alpha = .25
  ) +
  scale_fill_continuous(type = "viridis")

# with a zero radius or diameter
ggplot(df, aes(x = x, y = y)) +
  coord_fixed() +
  stat_simplicial_complex(radius = 0)
# with a too-small radius or diameter
ggplot(df, aes(x = x, y = y)) +
  coord_fixed() +
  stat_simplicial_complex(radius = 0.01)

# Visualizing multiple groups together
s <- c(s, s)
df_mix <- data.frame(
  x = cos(s) + rnorm(80, 0, .1),
  y = sin(s) + rnorm(80, 0, .1)
)
df_mix$x <- df_mix$x + rep(c(-2, 2), length.out = 80)
df_mix$lab <- rep(c("a", "b"), length.out = 80)

ggplot(df_mix, aes(x, y, fill = lab)) +
  geom_simplicial_complex(radius = .4)

# generate a noisy 2D circle
set.seed(2)
theta <- stats::runif(n = 40L, min = 0, max = 2*pi)
d <- data.frame(x = cos(theta) + stats::rnorm(40L, 0, .15),
                y = sin(theta) + stats::rnorm(40L, 0, .15))
r <- 1/3

# overlay ball cover and Vietoris-Rips complex with points
ggplot(d, aes(x = x, y = y)) +
  theme_bw() +
  coord_fixed() +
  geom_disk(radius = r, fill = "aquamarine3") +
  geom_simplicial_complex(
    radius = r, fill = "darkgoldenrod",
    complex = "Vietoris", engine = "base"
  ) +
  geom_point()
# use the Čech complex instead
ggplot(d, aes(x = x, y = y)) +
  theme_bw() +
  coord_fixed() +
  geom_disk(radius = r, fill = "aquamarine3") +
  geom_simplicial_complex(
    radius = r, fill = "darkgoldenrod",
    complex = "Cech", engine = "base"
  ) +
  geom_point()
}
\seealso{
\code{\link[ggplot2:layer]{ggplot2::layer()}} for additional arguments.

Other plot layers for point clouds: 
\code{\link{disk}}
}
\concept{plot layers for point clouds}
