% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist.R
\name{stat_flat}
\alias{stat_flat}
\title{Plot Persistent Homology as a Flat Persistence Diagram}
\usage{
stat_flat(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{geom}{The geometric object used to display the data.
Defaults to `point`; pass a string to override the default.}

\item{...}{additional arguments passed to [ggplot2::layer()]}
}
\description{
Plots a flat persistence diagram.
}
\seealso{
[ggplot2::layer()] for additional arguments.

Other TDA plot layers: \code{\link{geom_persist}}
}
\concept{TDA plot layers}
