% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_barcode}
\alias{theme_barcode}
\title{Theme for Persistence Barcodes}
\usage{
theme_barcode(vertical = FALSE)
}
\arguments{
\item{vertical}{defaults to \code{FALSE} (no vertical axis); if \code{TRUE}, allows
users to more easily count the number of features plotted in barcode
(if \code{vertical} equals \code{TRUE}, this function is equivalent to
\code{theme_persist})}
}
\description{
A simple theme to clearly display persistent homology in
the form of persistence barcodes (aka topological barcodes).
}
\seealso{
Other TDA themes: 
\code{\link{theme_persist}()}
}
\concept{TDA themes}
