% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_persist}
\alias{theme_persist}
\title{Theme for Persistence Diagrams}
\usage{
theme_persist()
}
\description{
A simple theme to clearly display persistent homology in
the form of flat persistence diagrams or diagonal persistence diagrams.
}
\seealso{
Other TDA themes: 
\code{\link{theme_barcode}()}
}
\concept{TDA themes}
