% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdatheme.R
\name{theme_tda}
\alias{theme_tda}
\title{Theme for Persistent Homology Visualizations}
\usage{
theme_tda()
}
\description{
A simple theme to clearly display persistent homology.
This minimalistic theme omits plot components that are
unnecessary in the context of persistent homology
(e.g. gridlines, background color).
This theme should get you close to a publication-quality
plot
}
\concept{TDA themes}
