% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{themes}
\alias{themes}
\alias{theme_persist}
\alias{theme_barcode}
\title{Themes for persistence data visualizations}
\usage{
theme_persist()

theme_barcode(vertical = FALSE)
}
\arguments{
\item{vertical}{Logical; whether to plot a vertical axis. Defaults to
\code{FALSE}; if \code{TRUE}, allows users to more easily count the number of
features plotted in barcode; equivalent to \code{theme_persist()}.}
}
\description{
A simple theme to clearly display persistent homology in the
form of persistence barcodes or diagrams.
}
