% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode.R
\name{stat_barcode}
\alias{stat_barcode}
\alias{geom_barcode}
\alias{barcode}
\title{Barcode Diagrams}
\usage{
stat_barcode(
  mapping = NULL,
  data = NULL,
  geom = "barcode",
  position = "identity",
  filtration = "Rips",
  diameter_max = NULL,
  radius_max = NULL,
  max_hom_degree = 1L,
  field_order = 2L,
  order_by = c("persistence", "start"),
  engine = NULL,
  infinity_break = Inf,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

geom_barcode(
  mapping = NULL,
  data = NULL,
  stat = "barcode",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[ggplot2:layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{filtration}{The type of filtration from which to compute persistent
homology; one of \code{"Rips"}, \code{"Vietoris"} (equivalent) or \code{"alpha"}.}

\item{diameter_max, radius_max}{Maximum diameter or radius for the simplicial
filtration. Both default to \code{NULL}, in which case the complete filtration
is constructed.}

\item{max_hom_degree}{Maximum dimension of the simplicial filtration,
the highest-dimensional features to be calculated.}

\item{field_order}{(Prime) order of the field over which to compute
persistent homology.}

\item{order_by}{A character vector comprised of (\code{"persistence"}, \code{"birth"}, and/or \code{"end"})
by which the features should be ordered (within \code{group});
defaults to \code{c("persistence", "birth")}.}

\item{engine}{The computational engine to use (see 'Details'). Reasonable
defaults are chosen based on \code{filtration}.}

\item{infinity_break}{Where to plot the death of features with infinite
persistence (i.e. \code{death == Inf}). Defaults to \code{Inf}, corresponding to the edge
of the plotting window.}

\item{na.rm}{Logical: if \code{FALSE}, the default, \code{NA} lodes are not included;
if \code{TRUE}, \code{NA} lodes constitute a separate category, plotted in grey
(regardless of the color scheme).}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}
}
\description{
Visualize persistence homologies with barcode diagrams.
}
\details{
\emph{Barcode diagrams} represent persistence data with \href{https://ggplot2.tidyverse.org/reference/geom_segment.html}{line segments}.

They can be understood as the "persistence analogue of a Betti number",
visualizing persistence homologies in such a way that significant topological features are made evident
even in the presence of noise (Ghrist, 2008).
}
\section{Persistence data}{


Persistence data encode the values of an underlying parameter
\eqn{\epsilon} at which topological features appear ("birth") and disappear
("death"). The difference between the birth and the death of a feature is
called its "persistence". As topological features may be of different
dimensions, persistence data sets usually also include the dimension of
each feature.

Persistence data can be specified in two ways in \strong{ggtda}:
\enumerate{
\item \strong{The \code{birth} and \code{death} aesthetics}. Pre-computed persistence homologies
can be supplied directly to layer functions in \strong{ggtda}, the columns of \code{data}
with the birth and death values for each feature must be mapped to the
\code{birth} and \code{death} aesthetics, respectively.
\item \strong{The \code{dataset} aesthetic}. Instead of a pre-computed persistence homology,
\strong{ggtda} will compute the persistence homologies for arbitrary data sets.
This requires the use of a nested \code{tibble} with list columns, the \code{data}
provided to either \code{ggplot()} or the layer function should be a \code{tibble}
with a column whose entries are data sets in a format that one of the
\strong{ggtda} engines supports. This list column must be mapped to the \code{dataset} aesthetic.
}

See the examples included below or
\code{vignette("grouped-list-data", package = "ggtda")} for an extended
discussion of the \code{dataset} aesthetic's use.
}

\section{Aesthetics}{

\code{stat_barcode()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{dataset} \emph{or} \code{birth}}
\item \strong{\code{dataset} \emph{or} \code{death}}
\item \code{group}
}
\code{geom_barcode()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \strong{\code{xend} \emph{or} \code{yend}}
\item \code{alpha}
\item \code{colour}
\item \code{group}
\item \code{linetype}
\item \code{linewidth}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs", package = "ggplot2")}.
}

\section{Computed variables}{

These are calculated by the 'stat' part of layers and can be accessed with \link[ggplot2:aes_eval]{delayed evaluation}.
\itemize{
\item \code{after_stat(dimension)}\cr integer feature dimension.
\item \code{after_stat(birth)}\cr birth value of each feature.
\item \code{after_stat(death)}\cr death value of each feature.
\item \code{after_stat(persistence)}\cr differences between birth and death values of features.
\item \code{after_stat(infinite)}\cr logical, is the feature's death at infinity
\item \code{after_stat(group)}\cr interaction of existing 'group', dataset ID, and 'dimension'.
}
}

\examples{

# toy example
toy.data <- data.frame(
  appear = c(0, 0, 0, 1, 2),
  disappear = c(5, 3, 3, 5, 3),
  dim = c("0", "0", "0", "2", "1")
)
# topological barcode using the geom layer (and minimalist theme)
ggplot(toy.data,
       aes(start = appear, end = disappear, colour = dim, shape = dim)) +
  geom_barcode() +
  theme_barcode()

# load library and dataset for comprehensive example
library("ripserr")
angles <- runif(100, 0, 2 * pi)
circle2d <- cbind(cos(angles), sin(angles)) # unit circle (Betti-1 number = 1)
# calculate persistence homology and format
circ.phom <- as.data.frame(vietoris_rips(circle2d))
circ.phom$dimension <- as.factor(circ.phom$dimension)
# pretty topological barcode with geom layer
ggplot(circ.phom, aes(start = birth, end = death,
                      colour = dimension)) +
  geom_barcode() +
  theme_barcode()
}
\references{
R Ghrist (2008) Barcodes: The persistent topology of data
\emph{Bulletin of the American Mathematical Society}, \strong{45} 61--75.
\url{https://ams.org/journals/bull/2008-45-01/S0273-0979-07-01191-3/S0273-0979-07-01191-3.pdf}

G Carlsson, A Zomorodian, A Collins, and L Guibas (2004) Persistence barcodes
for shapes. \emph{Proceedings of the 2004 Eurographics/ACM SIGGRAPH symposium on
Geometry processing}, 124--135. \doi{10.1145/1057432.1057449}

G Carlsson (2014) Topological pattern recognition for point cloud data. \emph{Acta
Numerica} \strong{23}, 289--368. \doi{10.1017/S0962492914000051}

F Chazal and B Michel (2017) An introduction to Topological Data Analysis:
fundamental and practical aspects for data scientists.
\url{https://arxiv.org/abs/1710.04019}
}
\seealso{
Other plot layers for persistence data: 
\code{\link{stat_landscape}()},
\code{\link{stat_persistence}()}
}
\concept{plot layers for persistence data}
