% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-key-simplex.R
\name{draw_key}
\alias{draw_key}
\alias{draw_key_simplex}
\title{Key glyphs for topological visualizations}
\usage{
draw_key_simplex(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\description{
The \code{simplex} key is equivalent to the \code{polygon} key except that
it renders triangles rather than squares. This evokes the triangularity
of simplices and distinguishes plots from plots of arbitrary polygons.
}
