% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplicial-complex.R
\name{geom_simplicial_complex}
\alias{geom_simplicial_complex}
\alias{stat_simplicial_complex}
\title{Simplicial complexes from 2-d point clouds}
\usage{
stat_simplicial_complex(
  mapping = NULL,
  data = NULL,
  geom = "SimplicialComplex",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  radius = NULL,
  diameter = NULL,
  zero_simplexes = "all",
  one_simplexes = "maximal",
  max_dimension = 10,
  complex = "Rips",
  engine = "simplextree",
  inherit.aes = TRUE,
  ...
)

geom_simplicial_complex(
  mapping = NULL,
  data = NULL,
  stat = "SimplicialComplex",
  position = "identity",
  outlines = TRUE,
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data, either as a
\code{ggproto} \code{Geom} subclass or as a string naming the geom stripped of the
\code{geom_} prefix (e.g. \code{"point"} rather than \code{"geom_point"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{radius, diameter}{The radius or diameter used to construct the simplicial complex.}

\item{zero_simplexes}{One of \code{"none"}, \code{"maximal"}, and \code{"all"} (default).}

\item{one_simplexes}{One of \code{"none"}, \code{"maximal"} (default), and \code{"all"}.}

\item{max_dimension}{Compute simplexes of dimension up to \code{max_dimension}
(only relevant for the Vietoris-Rips complex computed with the \code{simplextree} engine).}

\item{complex}{The type of complex to compute (either \code{"Vietoris"}, \code{"Rips"}, \code{"Cech"}, or \code{"alpha"}).}

\item{engine}{What computational engine to use. Reasonable defaults are chosen based on \code{complex}.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{outlines}{Should the outlines of polygons representing the simplexes be drawn?}
}
\description{
Desc here
}
\section{Aesthetics}{
 geom_hdr understands the following aesthetics (required
aesthetics are in bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item \strong{dim}
\item \strong{type}
\item \strong{simplex_id}
\item alpha
\item color
\item fill
\item group
\item linetype
\item size
}
}

\section{Computed variables}{


\describe{ \item{dim}{The dimension of the corresponding simplex, an ordered factor}}
\describe{ \item{type}{The type of simplex represented by each row, one of:
\code{"zero_simplex"}, \code{"one_simplex"}, \code{"k_simplex"}.}}
\describe{ \item{simplex_id}{A set of unique ids for each simplex, for each \code{type}}}
}

\examples{

set.seed(1)

s <- seq(0, 2*pi, length.out = 40)
df <- data.frame(
  x = cos(s) + rnorm(40, 0, .1),
  y = sin(s) + rnorm(40, 0, .1)
)

# default, visualizing dim w/ alpha:
ggplot(df, aes(x, y)) +
  geom_simplicial_complex(radius = .4)

# visualizing dim w/ fill:
ggplot(df, aes(x, y)) +
  geom_simplicial_complex(
    mapping = aes(fill = after_stat(dim)),
    alpha = .5, radius = .4
  )


# Visualizing multiple groups together
s <- c(s, s)
df_mix <- data.frame(
  x = cos(s) + rnorm(80, 0, .1),
  y = sin(s) + rnorm(80, 0, .1)
)

df_mix$x <- df_mix$x + rep(c(-2, 2), length.out = 80)
df_mix$lab <- rep(c("a", "b"), length.out = 80)

ggplot(df_mix, aes(x, y, fill = lab)) +
  geom_simplicial_complex(radius = .4)
}
