% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtda.R, R/persist.R
\docType{data}
\name{ggtda-ggproto}
\alias{ggtda-ggproto}
\alias{StatFlat}
\alias{GeomPersist}
\title{Base ggproto classes for ggtda}
\description{
Base ggproto classes for ggtda
}
\seealso{
[`ggplot2::ggplot2-ggproto`]
}
\keyword{datasets}
\keyword{internal}
