% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode.R, R/ggtda.R, R/persist.R,
%   R/simplicial-complex.R
\docType{data}
\name{GeomBarcode}
\alias{GeomBarcode}
\alias{ggtda-ggproto}
\alias{GeomPersist}
\alias{StatDisk}
\alias{StatVietoris1}
\alias{StatVietoris2}
\alias{StatCech2}
\title{Base \code{ggproto} classes for \emph{ggtda}}
\format{An object of class \code{GeomBarcode} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.}
\description{
Base \code{ggproto} classes for \emph{ggtda}
}
\seealso{
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::ggplot2-ggproto}}
}
\keyword{datasets}
\keyword{internal}
