% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode.R, R/disk.R, R/fundamental-box.R,
%   R/ggtda.R, R/persistence.R, R/landscape.R, R/simplicial-complex.R
\docType{data}
\name{GeomBarcode}
\alias{GeomBarcode}
\alias{GeomDisk}
\alias{GeomFundamentalBox}
\alias{ggtda-ggproto}
\alias{StatPersistence}
\alias{GeomPersistence}
\alias{GeomLandscape}
\alias{StatSimplicialComplex}
\alias{GeomSimplicialComplex}
\title{Base \code{ggproto} classes for \emph{ggtda}}
\format{
An object of class \code{GeomBarcode} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomPersistence} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\description{
Base \code{ggproto} classes for \emph{ggtda}
}
\seealso{
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::ggplot2-ggproto}}
}
\keyword{datasets}
\keyword{internal}
