% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode.R, R/ggtda.R, R/persistence.r,
%   R/simplicial-complex.R
\docType{data}
\name{GeomBarcode}
\alias{GeomBarcode}
\alias{ggtda-ggproto}
\alias{StatPersistence}
\alias{StatFrontier}
\alias{StatDisk}
\alias{StatVietoris0}
\alias{StatVietoris1}
\alias{StatVietoris2}
\alias{StatCech2}
\alias{GeomFace}
\title{Base \code{ggproto} classes for \emph{ggtda}}
\format{
An object of class \code{GeomBarcode} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatFrontier} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatDisk} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatVietoris0} (inherits from \code{StatIdentity}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.

An object of class \code{StatVietoris1} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatVietoris2} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatCech2} (inherits from \code{StatVietoris2}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.

An object of class \code{GeomFace} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\description{
Base \code{ggproto} classes for \emph{ggtda}
}
\seealso{
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::ggplot2-ggproto}}
}
\keyword{datasets}
\keyword{internal}
