% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode.R, R/ggtda.R, R/layer-disk.R,
%   R/persistence.r, R/simplicial-complex.R
\docType{data}
\name{GeomBarcode}
\alias{GeomBarcode}
\alias{ggtda-ggproto}
\alias{GeomDisk}
\alias{StatPersistence}
\alias{StatFrontier}
\alias{GeomFundamentalBox}
\alias{StatSimplicialComplex}
\alias{GeomSimplicialComplex}
\title{Base \code{ggproto} classes for \emph{ggtda}}
\format{
An object of class \code{GeomBarcode} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomDisk} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.

An object of class \code{StatFrontier} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatSimplicialComplex} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomSimplicialComplex} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\description{
Base \code{ggproto} classes for \emph{ggtda}
}
\seealso{
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::ggplot2-ggproto}}
}
\keyword{datasets}
\keyword{internal}
