% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode.R, R/ggtda.R, R/persist.R
\docType{data}
\name{GeomBarcode}
\alias{GeomBarcode}
\alias{ggtda-ggproto}
\alias{GeomPersist}
\title{Base ggproto classes for ggtda}
\format{An object of class \code{GeomBarcode} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.}
\description{
Base ggproto classes for ggtda
}
\seealso{
[`ggplot2::ggplot2-ggproto`]
}
\keyword{datasets}
\keyword{internal}
