% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtda.R
\docType{package}
\name{ggtda}
\alias{ggtda}
\alias{ggtda-package}
\title{A *ggplot2* Extension for Persistence Data}
\description{
*ggtda* is a user-friendly *ggplot2* extension for visualizing data arising
from topological data analysis pipelines, in particular persistence data.
Once persistent homology as been computed, *tidyverse* useRs can efficiently
visualize the persistence of features within the "layered grammar of
graphics" framework of *ggplot2*.
}
