% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtda.R
\docType{package}
\name{ggtda}
\alias{ggtda}
\title{A \emph{ggplot2} Extension for Persistence Data}
\description{
\emph{ggtda} is a user-friendly \emph{ggplot2} extension for visualizing data arising
from topological data analysis pipelines, in particular persistence data.
Once persistent homology as been computed, \emph{tidyverse} useRs can efficiently
visualize the persistence of features within the "layered grammar of
graphics" framework of \emph{ggplot2}.
}
