% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtda.R
\docType{package}
\name{ggtda}
\alias{ggtda}
\alias{ggtda-package}
\title{ggplot2-Compatible Visualization of Persistent Homology}
\description{
A user-friendly, ggplot2-compatible package to visualize
persistent homology to complement topological data analysis
pipelines. Once calculation of persistent homology is complete,
tidyverse useRs can now effectively visualize the persistent
homology within the existing grammar of graphics framework implemented
by ggplot2.
}
