% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{stat_landscape}
\alias{stat_landscape}
\alias{geom_landscape}
\alias{landscape}
\title{Persistence Landscapes}
\usage{
stat_landscape(
  mapping = NULL,
  data = NULL,
  geom = "landscape",
  position = "identity",
  filtration = "Rips",
  diameter_max = NULL,
  radius_max = NULL,
  max_hom_degree = 1L,
  field_order = 2L,
  engine = NULL,
  diagram = "landscape",
  n_levels = Inf,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

geom_landscape(
  mapping = NULL,
  data = NULL,
  stat = "landscape",
  position = "identity",
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[ggplot2:layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{filtration}{The type of filtration from which to compute persistent
homology; one of \code{"Rips"}, \code{"Vietoris"} (equivalent) or \code{"alpha"}.}

\item{diameter_max, radius_max}{Maximum diameter or radius for the simplicial
filtration. Both default to \code{NULL}, in which case the complete filtration
is constructed.}

\item{max_hom_degree}{Maximum dimension of the simplicial filtration,
the highest-dimensional features to be calculated.}

\item{field_order}{(Prime) order of the field over which to compute
persistent homology.}

\item{engine}{The computational engine to use (see 'Details'). Reasonable
defaults are chosen based on \code{filtration}.}

\item{diagram}{One of \code{"flat"}, \code{"diagonal"}, or \code{"landscape"}; the
orientation for the diagram should take.}

\item{n_levels}{The number of levels to compute and plot. If \code{Inf} (the
default), determined to be all levels.}

\item{na.rm}{Logical: if \code{FALSE}, the default, \code{NA} lodes are not included;
if \code{TRUE}, \code{NA} lodes constitute a separate category, plotted in grey
(regardless of the color scheme).}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{linemitre}{Line mitre limit (number greater than 1).}
}
\description{
Visualize persistence homologies with persistence landscapes.
}
\details{
\emph{Persistence landscapes}, anticipated by some alternative coordinatizations
of persistence diagrams, were proposed as Lipschitz functions that demarcate
the Pareto frontiers of persistence diagrams. They can be averaged over the
diagrams obtained from multiple data sets designed or hypothesized to have
been generated from the same underlying topological structure.

Persistence landscapes do not currently recognize extended persistence data.
}
\section{Persistence data}{


Persistence data encode the values of an underlying parameter
\eqn{\epsilon} at which topological features appear ("birth") and disappear
("death"). The difference between the birth and the death of a feature is
called its "persistence". As topological features may be of different
dimensions, persistence data sets usually also include the dimension of
each feature.

Persistence data can be specified in two ways in \strong{ggtda}:
\enumerate{
\item \strong{The \code{birth} and \code{death} aesthetics}. Pre-computed persistence homologies
can be supplied directly to layer functions in \strong{ggtda}, the columns of \code{data}
with the birth and death values for each feature must be mapped to the
\code{birth} and \code{death} aesthetics, respectively.
\item \strong{The \code{dataset} aesthetic}. Instead of a pre-computed persistence homology,
\strong{ggtda} will compute the persistence homologies for arbitrary data sets.
This requires the use of a nested \code{tibble} with list columns, the \code{data}
provided to either \code{ggplot()} or the layer function should be a \code{tibble}
with a column whose entries are data sets in a format that one of the
\strong{ggtda} engines supports. This list column must be mapped to the \code{dataset} aesthetic.
}

See the examples included below or
\code{vignette("grouped-list-data", package = "ggtda")} for an extended
discussion of the \code{dataset} aesthetic's use.
}

\section{Aesthetics}{

\code{stat_landscape()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{dataset} \emph{or} \code{birth}}
\item \strong{\code{dataset} \emph{or} \code{death}}
\item \code{group}
}
\code{geom_landscape()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \code{alpha}
\item \code{colour}
\item \code{group}
\item \code{linetype}
\item \code{linewidth}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs", package = "ggplot2")}.
}

\examples{
# toy example
toy.data <- data.frame(
  birth = c(0, 0, 1, 3, 4, 1.5),
  death = c(5, 3, 5, 4, 6, 3),
  dim = factor(c(0, 0, 1, 1, 2, 2))
)
# persistence diagram with landscape overlaid
ggplot(toy.data,
       aes(start = birth, end = death, colour = dim, shape = dim)) +
  theme_persist() +
  coord_equal() +
  stat_persistence() +
  stat_landscape(aes(alpha = -after_stat(level)), diagram = "diagonal") +
  lims(x = c(0, 8), y = c(0, NA)) +
  guides(alpha = "none")
# persistence landscape with diagram overlaid
ggplot(toy.data,
       aes(start = birth, end = death, colour = dim, shape = dim)) +
  theme_persist() +
  coord_equal() +
  stat_landscape(aes(linetype = after_stat(factor(level)))) +
  stat_persistence(diagram = "landscape") +
  lims(x = c(0, 8), y = c(0, NA)) +
  labs(linetype = "level")

# load library and generate dataset for comprehensive example
library("ripserr")
# noisy unit circle (Betti-1 number = 1)
n <- 100L; sd <- 0.1
set.seed(7)
t <- stats::runif(n = n, min = 0, max = 2*pi)
annulus.df <- data.frame(
  x = cos(t) + stats::rnorm(n = n, mean = 0, sd = sd),
  y = sin(t) + stats::rnorm(n = n, mean = 0, sd = sd)
)
# calculate persistence homology and format
annulus.phom <- as.data.frame(vietoris_rips(annulus.df))
annulus.phom$dimension <- as.factor(annulus.phom$dimension)
# pretty diagonal persistence diagram
ggplot(annulus.phom, aes(start = birth, end = death,
                         shape = dimension, colour = dimension)) +
  stat_persistence(diagram = "landscape") +
  theme_persist()
# pretty landscape persistence diagram
ggplot(annulus.phom, aes(start = birth, end = death,
                         shape = dimension, colour = dimension)) +
  stat_landscape(diagram = "landscape") +
  theme_persist()
}
\references{
P Bubenik (2015) Statistical Topological Data Analysis using Persistence
Landscapes. \emph{Journal of Machine Learning Research}, \strong{16} 77--102.
\url{http://jmlr.org/papers/v16/bubenik15a.html}

F Chazal and B Michel (2017) An introduction to Topological Data Analysis:
fundamental and practical aspects for data scientists.
\url{https://arxiv.org/abs/1710.04019}
}
\seealso{
Other plot layers for persistence data: 
\code{\link{stat_barcode}()},
\code{\link{stat_persistence}()}
}
\concept{plot layers for persistence data}
