% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.r
\name{persistence}
\alias{persistence}
\alias{stat_persistence}
\alias{stat_frontier}
\title{Persistence diagrams}
\usage{
stat_persistence(mapping = NULL, data = NULL, geom = "point",
  position = "identity", diagram = "flat", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)

stat_frontier(mapping = NULL, data = NULL, geom = "line",
  position = "identity", diagram = "flat", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data; defaults to
\code{segment} in \code{geom_vietoris1()} and to \code{polygon} in \code{geom_vietoris2}. Pass
a string to override the default.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{diagram}{One of \code{"flat"}, \code{"diagonal"}, or \code{"landscape"}; the
orientation for the diagram should take}

\item{na.rm}{Logical: if \code{FALSE}, the default, \code{NA} lodes are not included;
if \code{TRUE}, \code{NA} lodes constitute a separate category, plotted in grey
(regardless of the color scheme).}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}
}
\description{
Visualize persistence data in a (flat, diagonal, or landscape)
persistence diagram.
}
\details{
\emph{Persistence diagrams} are
\href{https://ggplot2.tidyverse.org/reference/geom_point.html}{scatterplots} of
persistence data. \emph{Persistence landscapes} can be understood as rotated
diagonal persistence diagrams.
}
\section{Persistence data}{


\emph{Persistence data} encode the values of an underlying parameter
\eqn{\epsilon} at which topological features appear ("birth") and disappear
("death"). The difference between the birth and the death of a feature is
called its \emph{persistence}. Whereas topological features may be of different
dimensions, persistence data sets usually also include the dimension of
each feature.

\strong{ggtda} expects persistence data to have at least three columns: birth,
death, and dimension.
}

\section{Persistence diagrams}{


The original persistence diagrams plotted persistence against birth in what
we call "flat" diagrams, but most plot death against birth in "diagonal"
diagrams, often with a diagonal line indicating zero persistence.
}

\section{Persistence landscapes}{


Persistence landscapes, anticipated by some alternative coordinatizations
of persistence diagrams, were proposed as Lipschitz functions that
demarcate the Pareto frontiers of persistence diagrams. They can be
averaged over the diagrams obtained from multiple data sets designed or
hypothesized to have been generated from the same underlying topological
structure.
}

\examples{
#####EXAMPLE 1#####
# toy example
toy.data <- data.frame(
  birth = c(0, 0, 1, 2, 1.5),
  death = c(5, 3, 5, 3, 6),
  dim = c("0", "0", "2", "1", "1")
)
# flat persistence diagram
ggplot(toy.data,
       aes(start = birth, end = death, colour = dim, shape = dim)) +
  theme_tda() +
  stat_persistence() +
  lims(x = c(0, NA), y = c(0, NA))
# diagonal persistence diagram with frontier
ggplot(toy.data,
       aes(start = birth, end = death, colour = dim, shape = dim)) +
  theme_tda() + coord_equal() +
  stat_persistence(diagram = "diagonal") +
  stat_frontier(diagram = "diagonal") +
  lims(x = c(0, NA), y = c(0, NA))
# landscape persistence frontier
ggplot(toy.data,
       aes(start = birth, end = death, colour = dim, shape = dim)) +
  theme_tda() + coord_equal() +
  stat_frontier(diagram = "landscape") +
  lims(x = c(0, NA), y = c(0, NA))

#####EXAMPLE 2#####
# load library and dataset for comprehensive example
library("TDAstats")
data("annulus2d")  # noisy unit circle (Betti-1 number = 1)
# calculate persistence homology and format
annulus.phom <- as.data.frame(calculate_homology(annulus2d))
annulus.phom$dimension <- as.factor(annulus.phom$dimension)
# pretty flat persistence diagram
ggplot(annulus.phom, aes(start = birth, end = death,
                         shape = dimension, colour = dimension)) +
  stat_persistence(diagram = "flat") +
  theme_persist()
# pretty diagonal persistence diagram
ggplot(annulus.phom, aes(start = birth, end = death,
                         shape = dimension, colour = dimension)) +
  stat_persistence(diagram = "diagonal") +
  theme_persist()
# pretty landscape persistence diagram
ggplot(annulus.phom, aes(start = birth, end = death,
                         shape = dimension, colour = dimension)) +
  stat_frontier(diagram = "landscape") +
  theme_persist()
}
\references{
H Edelsbrunner, D Letscher, and A Zomorodian (2000) Topological persistence
and simplification. \emph{Proceedings 41st Annual Symposium on Foundations of
Computer Science}, 454--463. \doi{10.1109/SFCS.2000.892133}

H Edelsbrunner and D Morozov (2012) Persistent Homology: Theory and Practice.
\emph{European Congress of Mathematics}, 31--50. \doi{10.4171/120}

P Bubenik (2015) Statistical Topological Data Analysis using Persistence
Landscapes. \emph{Journal of Machine Learning Research}, \strong{16} 77--102.
\url{http://jmlr.org/papers/v16/bubenik15a.html}

F Chazal and B Michel (2017) An introduction to Topological Data Analysis:
fundamental and practical aspects for data scientists.
\url{https://arxiv.org/abs/1710.04019}
}
\seealso{
\code{\link[ggplot2:layer]{ggplot2::layer()}} for additional arguments.

Other plot layers for persistence data: \code{\link{barcode}}
}
\concept{plot layers for persistence data}
