% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist.R
\name{stat_flat}
\alias{stat_flat}
\title{Plot Persistent Homology as a Flat Persistence Diagram}
\usage{
stat_flat(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data.}

\item{geom}{The geometric object used to display the data.
Defaults to `point`; pass a string to override the default.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{Logical:
if `FALSE`, the default, `NA` values are not included
if `TRUE`, `NA` values constitute a separate category,
plotted in grey (regardless of the color scheme)}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{additional arguments passed to [ggplot2::layer()]}
}
\description{
Plots a flat persistence diagram.
}
\seealso{
[ggplot2::layer()] for additional arguments.

Other TDA plot layers: \code{\link{barcode}},
  \code{\link{geom_persist}}
}
\concept{TDA plot layers}
