% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_barcode}
\alias{theme_barcode}
\title{Theme for Persistence Barcodes}
\usage{
theme_barcode()
}
\description{
A simple theme to clearly display persistent homology in
the form of persistence barcodes (aka topological barcodes).
}
\seealso{
Other TDA themes: 
\code{\link{theme_persist}()},
\code{\link{theme_tda}()}
}
\concept{TDA themes}
